+++ src/data/services/spending-impact.test.ts
