+++ src/data/services/spending-impact.ts
