+++ src/routes/api/v1/fuel-type-usage/index.csv.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from '@sveltejs/kit';
 
-import { getFuelUsageDataFrame } from '$data/transforms/fuel-usage';
+import { getFuelUsageDataFrame } from '$data/services/fuel-usage';
 
 export const get: RequestHandler = async () => {
   const fuelUsage = await getFuelUsageDataFrame();
