+++ src/routes/api/v1/fuel-type-usage/index.json.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from '@sveltejs/kit';
 
-import { getFuelUsage } from '$data/transforms/fuel-usage';
+import { getFuelUsage } from '$data/services/fuel-usage';
 
 export const get: RequestHandler = async () => {
   const fuelUsage = await getFuelUsage();
